AI_WAWA/requirements.txt
# ===========================================
# AI 언어학습 앱 의존성 - 충돌 해결 버전
# ===========================================

# ===========================================
# 핵심 웹 프레임워크
# ===========================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# ===========================================
# 데이터 모델 및 설정
# ===========================================
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# ===========================================
# OpenAI 서비스 (호환 버전으로 조정)
# ===========================================
openai>=1.6.1
# langchain 제거 (실제 코드에서 사용하지 않음)
# torch/transformers는 실제 필요시에만 설치하도록 주석 처리
# transformers==4.35.0
# torch==2.1.0
# torchaudio==2.1.0

# ===========================================
# 음성 처리 (STT/TTS)
# ===========================================
SpeechRecognition==3.10.0
pyttsx3==2.90
soundfile==0.12.1

# ===========================================
# 고급 음성 분석 (발음 평가) - 선택적
# ===========================================
librosa==0.10.1
numpy==1.24.3
scipy==1.11.4
praat-parselmouth==0.4.3

# ===========================================
# 번역 및 언어 처리 (충돌 해결)
# ===========================================
deep-translator==1.11.4
langdetect==1.0.9
# language-tool-python은 Java 의존성이 필요해서 주석 처리
# language-tool-python==2.7.1

# ===========================================
# 데이터 수집 및 웹 스크래핑 (버전 조정)
# ===========================================
aiohttp==3.9.1
requests>=2.32.0
beautifulsoup4==4.12.2
praw==7.7.1

# ===========================================
# 데이터 처리 및 분석
# ===========================================
pandas==2.1.4
scikit-learn==1.3.0
textdistance==4.6.0

# ===========================================
# 데이터베이스 (선택적)
# ===========================================
sqlalchemy==2.0.23
alembic==1.12.1
supabase

# ===========================================
# 로깅
# ===========================================
loguru==0.7.2

# ===========================================
# 테스팅
# ===========================================
pytest==7.4.3
pytest-asyncio==0.21.1
httpx>=0.27.0

# ===========================================
# 유틸리티 라이브러리 (충돌 해결)
# ===========================================
# asyncio-mqtt는 실제 사용하지 않으므로 제거
phonemizer==3.2.1

# ===========================================
# 추가 설치
# ===========================================
elevenlabs==0.2.26
aiofiles==23.2.1
datasets

# ===========================================
# 추가 고급 기능 (필요시 개별 설치)
# ===========================================
# 다음 패키지들은 필요한 경우에만 개별 설치
# pip install torch torchaudio transformers  # AI 모델 훈련 필요시
# pip install sounddevice  # 마이크 직접 접근 필요시
# pip install language-tool-python  # 문법 검사 필요시
# pip install asyncio-mqtt  # MQTT 통신 필요시
# pip install psycopg2-binary  # PostgreSQL 사용시
# pip install redis  # Redis 캐싱 사용시
